# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "127.0.0.1:5555"
basePath: "/api/v1"
schemes:
- "http"
paths:
  /g-demo:
    get:
      tags:
      - "g-demo"
      summary: "get all g-demo by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GDemoPagination"
    post:
      tags:
      - "g-demo"
      summary: "create g-demo"
      description: "create g-demo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-demo"
        required: true
        schema:
          $ref: "#/definitions/GDemo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-demo"
      summary: "update g-demo"
      description: "update g-demo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-demo"
        required: true
        schema:
          $ref: "#/definitions/GDemo"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-demo/{Id}:
    get:
      tags:
      - "g-demo"
      summary: "get a g-demo by ID"
      description: "get a g-demo by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-demo to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GDemo"
    delete:
      tags:
      - "g-demo"
      summary: "Destroy a g-demo by ID"
      description: "delete a g-demo by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-demo to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-menu:
    get:
      tags:
      - "g-menu"
      summary: "get all g-menu by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GMenuPagination"
    post:
      tags:
      - "g-menu"
      summary: "create g-menu"
      description: "create g-menu"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-menu"
        required: true
        schema:
          $ref: "#/definitions/GMenu"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-menu"
      summary: "update g-menu"
      description: "update g-menu"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-menu"
        required: true
        schema:
          $ref: "#/definitions/GMenu"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-menu/{Id}:
    get:
      tags:
      - "g-menu"
      summary: "get a g-menu by ID"
      description: "get a g-menu by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-menu to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GMenu"
    delete:
      tags:
      - "g-menu"
      summary: "Destroy a g-menu by ID"
      description: "delete a g-menu by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-menu to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-menu-action:
    get:
      tags:
      - "g-menu-action"
      summary: "get all g-menu-action by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GMenuActionPagination"
    post:
      tags:
      - "g-menu-action"
      summary: "create g-menu-action"
      description: "create g-menu-action"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-menu-action"
        required: true
        schema:
          $ref: "#/definitions/GMenuAction"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-menu-action"
      summary: "update g-menu-action"
      description: "update g-menu-action"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-menu-action"
        required: true
        schema:
          $ref: "#/definitions/GMenuAction"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-menu-action/{Id}:
    get:
      tags:
      - "g-menu-action"
      summary: "get a g-menu-action by ID"
      description: "get a g-menu-action by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-menu-action to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GMenuAction"
    delete:
      tags:
      - "g-menu-action"
      summary: "Destroy a g-menu-action by ID"
      description: "delete a g-menu-action by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-menu-action to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-menu-resource:
    get:
      tags:
      - "g-menu-resource"
      summary: "get all g-menu-resource by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GMenuResourcePagination"
    post:
      tags:
      - "g-menu-resource"
      summary: "create g-menu-resource"
      description: "create g-menu-resource"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-menu-resource"
        required: true
        schema:
          $ref: "#/definitions/GMenuResource"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "g-menu-resource"
      summary: "update g-menu-resource"
      description: "update g-menu-resource"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create g-menu-resource"
        required: true
        schema:
          $ref: "#/definitions/GMenuResource"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /g-menu-resource/{Id}:
    get:
      tags:
      - "g-menu-resource"
      summary: "get a g-menu-resource by ID"
      description: "get a g-menu-resource by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-menu-resource to return"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GMenuResource"
    delete:
      tags:
      - "g-menu-resource"
      summary: "Destroy a g-menu-resource by ID"
      description: "delete a g-menu-resource by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of g-menu-resource to delete"
        required: true
        type:  "integer" 
        format:  "int64"  
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
definitions:
  GDemo:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      record_id:
        type: "string"
        description: ""
        format: "string"
      
      code:
        type: "string"
        description: ""
        format: "string"
      
      name:
        type: "string"
        description: ""
        format: "string"
      
      memo:
        type: "string"
        description: ""
        format: "string"
      
      status:
        type: "integer"
        description: ""
        format: "int64"
      
      creator:
        type: "string"
        description: ""
        format: "string"
  GDemoPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GDemo"
  GMenu:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      record_id:
        type: "string"
        description: ""
        format: "string"
      
      name:
        type: "string"
        description: ""
        format: "string"
      
      sequence:
        type: "integer"
        description: ""
        format: "int64"
      
      icon:
        type: "string"
        description: ""
        format: "string"
      
      router:
        type: "string"
        description: ""
        format: "string"
      
      hidden:
        type: "integer"
        description: ""
        format: "int64"
      
      parent_id:
        type: "string"
        description: ""
        format: "string"
      
      parent_path:
        type: "string"
        description: ""
        format: "string"
      
      creator:
        type: "string"
        description: ""
        format: "string"
  GMenuPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GMenu"
  GMenuAction:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      menu_id:
        type: "string"
        description: ""
        format: "string"
      
      code:
        type: "string"
        description: ""
        format: "string"
      
      name:
        type: "string"
        description: ""
        format: "string"
  GMenuActionPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GMenuAction"
  GMenuResource:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: ""
        format: "int32"
      
      created_at:
        type: "string"
        description: ""
        format: "date-time"
      
      updated_at:
        type: "string"
        description: ""
        format: "date-time"
      
      deleted_at:
        type: "string"
        description: ""
        format: "date-time"
      
      menu_id:
        type: "string"
        description: ""
        format: "string"
      
      code:
        type: "string"
        description: ""
        format: "string"
      
      name:
        type: "string"
        description: ""
        format: "string"
      
      method:
        type: "string"
        description: ""
        format: "string"
      
      path:
        type: "string"
        description: ""
        format: "string"
  GMenuResourcePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/GMenuResource"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...